Fix loading of arguments so that defaults are actually used
There is a bug where the default args are not being used. I believe the issue is that the input args exist, but are empty. To be honest, I'm not sure if this is an artifact of python that I don't understand or if its a Github Action oddity. In either case, I'm hoping this fixes the issue by using the defaults if the arg is empty.