Allow passing of good and bad examples
Chat GPT seems to be a bit lost without some concrete examples for each rule, so I updated the input to accept an array of rules with a description as well as good/bad examples. Additionally, I updated the prompt to describe the rule format. Because of this, I made it so that the entire prompt can not be overridden. Instead, it can be customized with additional notes. This will ensure that the instructions regarding rule format are included.

Unfortunately, Github Actions do not support arrays for inputs at the moment, so I had to implement this as a string. A JSON array felt the easiest for communicating to ChatGPT as it is unlikely to have linebreak issues or conflicts with other symbols like "-" which can cause confusion when breaking apart a list.